+++ src/app/core/services/adapters/meta-content.adapter.ts
@@ -3,6 +3,7 @@ import {Observable} from "rxjs";
 import {map} from "rxjs/operators";
 import {MetaMapperData} from "@app/modules/pages/editor/models";
 import {EditorPreviewInfoService} from "@app/core/model/content.service";
+import {environment} from "@environment/environment";
 
 export interface MetaData {
   metaDescription: string;
@@ -23,15 +24,14 @@ export interface MetaData {
  * Downloads meta data of certain given website by internal api
  */
 export class MetaContentAdapter implements EditorPreviewInfoService {
-  private readonly apiUrl: string;
+  private readonly META_ENDPOINT: string = `${environment.API_URL}/meta`;
 
   constructor(private httpClient: HttpClient) {
-    this.apiUrl = '/api/meta';
   }
 
   getMetaMapperData(url: string): Observable<MetaMapperData> {
     const b64url = btoa(url);
-    return this.httpClient.get<MetaData>(`${this.apiUrl}/${b64url}`).pipe(
+    return this.httpClient.get<MetaData>(`${this.META_ENDPOINT}/${b64url}`).pipe(
       map(data => ({
         title: data.title,
         description: data.metaDescription,