+++ src/app/core/services/cookie.service.ts
@@ -1,5 +1,6 @@
 import {Injectable} from '@angular/core';
 import {SimpleLocalCacheService} from "@app/core/services/simple-local-cache.service";
+import {environment} from "@environment/environment";
 
 @Injectable({
   providedIn: 'root'
@@ -9,6 +10,10 @@ export class CookieService {
   }
 
   isTrackingDisabled(): boolean {
+    if (!environment.production) {
+      return true;
+    }
+
     const enabled = this.simpleLocalCache.getItem('cookies');
     return enabled === 'disabled';
   }