+++ src/environments/environment.qa.ts
@@ -1,0 +1,5 @@
+export const environment = {
+  production: false,
+  API_URL: 'https://www.meta-mapper.com/api',
+  CMS_URL: 'https://www.meta-mapper.com/cms'
+};